lili_app_documentado/requirements.txt
streamlit
pandas
numpy
python-docx
sounddevice
vosk
fpdf
watchdog
gpt4all
pyttsx3
tinydb